requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
idna==2.8
jira==2.0.0
oauthlib==3.1.0
pbr==5.4.2
pycparser==2.19
PyJWT==1.7.1
python-dotenv==0.10.3
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
six==1.12.0
urllib3==1.25.3
